refactor(stats): improve generic programming with the `Stats` trait
- `Stats` is implemented for `[T]` where `T: Simd`. Since `Simd` is implemented
  by `f32`, `f64`, the `T: Simd` bound can be used for generic programming
- the slow `::t` has been removed, use `Stats::t` instead
- T tests (`TDistribution`) are now SIMD accelerated

Closes #12
Closes #13

[breaking-change]
